+++ src/__tests__/offerwall.test.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import * as api from '../api/bitlabs_api';
+import * as api from '../api/bitlabs/bitlabs_api';
 import { act, render, screen, waitFor } from '@testing-library/react-native';
 import '@testing-library/jest-native/extend-expect';
 import OfferWall from '../components/offerwall';
 import { Platform } from 'react-native';
-import { HookName, type HookMessage } from '../api/types';
+import { HookName, type HookMessage } from '../api/bitlabs/types';
 
 const mockProps = {
   token: 'token',
