+++ src/components/fullwidth-survey.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Image, Text, TouchableOpacity, View } from 'react-native';
-import type { Survey, SurveyProperties } from '../api/types';
+import type { Survey, SurveyProperties } from '../api/bitlabs/types';
 import images from '../assets/images';
 import SurveyStyles from '../styles/fullwidth-survey.styles';
 import RatingBar from '../hoc/rating-bar';
