+++ src/components/survey-list.tsx
@@ -2,12 +2,16 @@ import { FlatList, type StyleProp, type ViewStyle } from 'react-native';
 import React, { useEffect, useState } from 'react';
 import SimpleWidget from './simple-survey';
 import CompactWidget from './compact-survey';
-import { type Survey, type SurveyProperties, WidgetType } from '../api/types';
+import {
+  type Survey,
+  type SurveyProperties,
+  WidgetType,
+} from '../api/bitlabs/types';
 import {
   getAppSettings,
   getIsImageSVG,
   getSurveysRepo,
-} from '../api/bitlabs_repository';
+} from '../api/bitlabs/bitlabs_repository';
 import FullWidthWidget from './fullwidth-survey';
 import { extractColors } from '../utils/helpers';
 import Gradient from '../hoc/gradient';
